requirements.txt
numpy
opencv-python
pytorch
torchvision